+++ CHANGES.md
@@ -3,6 +3,7 @@
 [full changelog](http://github.com/yolk/valvat/compare/v0.1.1...master)
 
 * Rewrote Valvat module to a vat number class for convenience + internal use of Valvat instances
+* I18n: Default error messages in german
 
 ### 0.1.1 / 2011-01-07